+++ src/Frameworks/Laravel/ServiceProvider.php
@@ -108,6 +108,8 @@ class ServiceProvider extends IlluminateServiceProvider
             \AvtoDev\FakerProviders\Providers\Packages\IDEntityProvider::class,
 
             \AvtoDev\FakerProviders\Providers\User\AvatarUriProvider::class,
+
+            \AvtoDev\FakerProviders\Providers\Identifiers\InnAndKppProvider::class,
         ];
     }
 }