+++ tests/Frameworks/Laravel/ServiceProviderTest.php
@@ -47,6 +47,8 @@ class ServiceProviderTest extends IlluminateTestCase
 
             \AvtoDev\FakerProviders\Providers\User\AvatarUriProvider::class,
 
+            \AvtoDev\FakerProviders\Providers\Identifiers\InnAndKppProvider::class,
+
         ];
 
         foreach ($providers as $provider_class) {